moved testDataHelper to helper folder
