+++ integration_tests/tests/e2e-tests.ts
@@ -19,7 +19,7 @@ import {
   retrieveMoreTestDataFromConfig,
   TestData,
   TestDataRetrievedFromConfig,
-} from "../../src/handlers/int-test-support/testDataHelper";
+} from "../../src/handlers/int-test-support/helpers/testDataHelper";
 import { BillingTransactionCurated } from "./billing-and-transaction-view-tests";
 
 const prefix = resourcePrefix();
